Timer and Dual Temp
Adds a timer function to set the cooking duration, for Issue #6 and also
a dual temperature for Issue #7
Functionality works as follows:
While in RUN mode, the SELECT button starts the timer.
While timer is running, the target temperature will be SP1
When timer is finished, the target temperature will be SP2

This allows you to set SP2 to a 'keep warm temperature', which will let
machine warm up to this temperature before you start the timer and after
the timer is finished it will return to this keep warm temperature.